Update Dutch database GK069
It should be the last voyage of the ship Prins Willem de vijfde. Tstd indeed says "unknown outcome". In fact the ship all but perished by a hurricane of the Caribbean island of St Eustatius on 31 August 1772. Slaves were already disembarked. What remained of the ship was unfit for a voyage back and was sold, insurers also paid out.

Signed-off-by: Guillaume Daudin <gdaudin@mac.com>